Refactored parser function 'parse_gen_stmt'
The old function, 'parse_gen_stmt', did not have a good name and was
needlessly complicated (returned a function for no reason). To fix this
issue, the function was renamed to 'consume_semicolon' which makes its
purpose much more clear and was changed to a currying funcion. This also
meant that the call sites could be simplified to simply call the
function instead of generating a function and then imidiatly calling it.