Added all assignment integration tests
The original lox repository includes a large suite of integration tests.
This change adds the assignment directory of integration tests and fixes
any bugs found along the way. This included a bug with how comments were
parsed (we now skip comments instead of creating token for them) and a
bug with printing 'nil' at the end of running a script (this is still a
feature for the REPL). In the process, I also started the process of
creating an opam specification for the process.