Added block, bool, call, class integration tests
Added 4 more of the integration test suites from the main lox
repository. The main bug found and fix is the inability for the parser
to handle extra semicolons. This was fixed and there is a new unit test
for this case. At the same time, some of the error reporting in unit
tests was improved.